Added Post Creation and File Upload Code - Raymond Muller
Rough draft here.  Awaiting confirmation on the exact mechanics of coordinating front end and back end.  Specifically, I need to ensure that the POST request in PostCreateComponent.php (lines 58-80) is correct, and I need to link it to the controller code.

Further, the upload file mechanic needs to be fleshed out.  Currently, files are stored in the /upload folder.  Once a mechanism for storing images and video at the back end is determined, the Upload.php code and the Post handling code in PostCreateComponent need to be modified to forward the uploaded video.


Finally, since this is a subcomponent, I made some assumptions about the session variables.  Namely, I retrieved username and postCount through $_SESSION['varName'].  This could change depending on back-end implementation, and also depending on how the profile info component and profile page are implemented.